adding explaining comment for dataSave methods
add 2 method to dataSave object: read and readDataThenInsert. read is for reading the store properties in localStorage. readDataThenInsert is for reading the user inputted data and insert it into localStorage

add a new section for transfering data between devices. adding a button for coping the stored data in the current device, and another button for inserting the copied data from another device

change some logic for changing edit.classBusIndexTracker in the change()

add dataSave.push() for deleting data, dunno why it not there before

add a function for copying texts to clipboard

fix Cancle typo